restructure web-pages
