+++ CONTRIBUTING.md
@@ -52,14 +52,12 @@ The marketing pages for the application. They are generated by a rust program wh
 
 ## Folder: web-pages
 
-- Every route in the application has a corresponding page. 
-- For example if we have a route `/api_keys` then we will have an `api_keys.rs`.
-- This will have a function that `index` that takes the parameteres needed to populate the page.
-- The `index` function uses rsx! to generate the page using.
-- If the page gets large we can break it up into components.
-- If the file gets large we may break the file into a module.
-- So for example `api_keys.rs` would have a corresponding folder called `api_keys` with components inside it.
-- We use Tailwind and Daisy UI. Only use daisy UI colors and when possible try to use the provided Daisy RSX library.
+- Every route has its own folder under `crates/web-pages`.
+- The main page for a route lives in a file called `page.rs` inside that folder.
+- Additional components are stored either alongside `page.rs` or in a `components/` folder.
+- Shared widgets such as confirmation dialogs live under `components/` at the crate root.
+- Each page corresponds to a typed route defined in `crates/web-pages/routes.rs` and is called from the matching handler in `crates/web-server/handlers`.
+- We use Tailwind and Daisy UI. Only use Daisy UI colors and when possible the provided Daisy RSX library.
 
 ## Folder: web-server
