+++ crates/web-pages/api_keys/mod.rs
@@ -1,2 +1,2 @@
 pub mod form;
-pub mod index;
+pub mod page;
