+++ crates/web-pages/api_keys/page.rs
@@ -2,8 +2,9 @@
 use crate::{
     app_layout::{Layout, SideBar},
     charts::{ApiRequestChartCard, TokenUsageChartCard},
-    render, ConfirmModal,
+    render,
 };
+use crate::components::confirm_modal::ConfirmModal;
 use assets::files::*;
 use daisy_rsx::*;
 use db::{authz::Rbac, ApiKey, Prompt, PromptType as DBPromptType};
