+++ crates/web-pages/assistants/mod.rs
@@ -1,6 +1,6 @@
 pub mod assistant_console;
 pub mod conversation;
-pub mod index;
+pub mod page;
 pub mod prompt_card;
 pub mod view_prompt;
 pub mod visibility;
