+++ crates/web-pages/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod confirm_modal;
+pub mod logout_form;
