+++ crates/web-pages/console/conversation.rs
@@ -3,7 +3,7 @@
 use super::{ChatWithChunks, PendingChatState};
 use crate::app_layout::SideBar;
 use crate::console::model_popup::ModelPopup;
-use crate::ConfirmModal;
+use crate::components::confirm_modal::ConfirmModal;
 use assets::files::*;
 use daisy_rsx::*;
 use db::authz::Rbac;
