+++ crates/web-pages/console/mod.rs
@@ -2,7 +2,7 @@ pub mod console_stream;
 pub mod conversation;
 pub mod empty_stream;
 pub mod history_drawer;
-pub mod index;
+pub mod page;
 pub mod layout;
 pub mod model_popup;
 pub mod prompt_drawer;
