+++ crates/web-pages/datasets/mod.rs
@@ -1,2 +1,2 @@
-pub mod index;
+pub mod page;
 pub mod upsert;
