+++ crates/web-pages/documents/mod.rs
@@ -1,4 +1,4 @@
-pub mod index;
+pub mod page;
 pub mod status;
 pub mod upload;
