+++ crates/web-pages/integrations/integration_header.rs
@@ -1,6 +1,6 @@
 #![allow(non_snake_case)]
 use crate::routes;
-use crate::ConfirmModal;
+use crate::components::confirm_modal::ConfirmModal;
 use assets::files::{button_edit_svg, menu_delete_svg};
 use daisy_rsx::*;
 use db::{authz::Rbac, Integration};
