+++ crates/web-pages/integrations/mod.rs
@@ -2,7 +2,7 @@ pub mod actions_section;
 pub mod api_key_cards;
 pub mod api_key_form;
 pub mod connections_section;
-pub mod index;
+pub mod page;
 pub mod integration_cards;
 pub mod integration_header;
 pub mod integration_type;
