+++ crates/web-pages/lib.rs
@@ -8,16 +8,14 @@ pub mod audit_trail;
 pub mod automations;
 pub mod base_layout;
 pub mod charts;
-pub mod confirm_modal;
+pub mod components;
 pub mod console;
 pub mod datasets;
 pub mod my_assistants;
-pub use confirm_modal::ConfirmModal;
 pub mod documents;
 pub mod hero;
 pub mod history;
 pub mod integrations;
-pub mod logout_form;
 pub mod menu;
 pub mod models;
 pub mod oauth_clients;
