+++ crates/web-pages/models/mod.rs
@@ -1,4 +1,4 @@
-pub mod index;
+pub mod page;
 pub mod model_card;
 pub mod model_type;
 pub mod upsert;
