+++ crates/web-pages/my_assistants/mod.rs
@@ -1,6 +1,6 @@
 pub mod assistant_card;
 pub mod connection_modal;
 pub mod datasets;
-pub mod index;
+pub mod page;
 pub mod integrations;
 pub mod upsert;
