+++ crates/web-pages/oauth_clients/mod.rs
@@ -1,3 +1,3 @@
-pub mod index;
+pub mod page;
 pub mod oauth_client_card;
 pub mod upsert;
