+++ crates/web-pages/pipelines/mod.rs
@@ -1,2 +1,2 @@
-pub mod index;
+pub mod page;
 pub mod key_drawer;
