+++ crates/web-pages/teams/mod.rs
@@ -1,4 +1,4 @@
 pub mod accept_invitation;
-pub mod index;
+pub mod page;
 pub mod invite_card;
 pub mod team_card;
